+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Improve testing:
   - Set client_session to None
+  - Remove unnecessary async from create_central
 
 # Version 2023.8.4 (2023-08-14)